separate flightbuilder from momondo scrapper
